autenticacao api catalogo e projeto auth
