+++ RafaStore.WebAPI.Core/RafaStore.WebAPI.Core.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="microsoft.aspnetcore.authentication.jwtbearer" Version="8.0.8" />
+	</ItemGroup>
+
+</Project>
