+++ RafaStore.sln
@@ -25,7 +25,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RafaStore.WebApp.MVC", "src
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Catalogo", "Catalogo", "{8C652B01-5A8B-4FB4-9C29-791A4ACC3105}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RafaStore.Catalogo.API", "src\services\RafaStore.Catalogo.API\RafaStore.Catalogo.API.csproj", "{184DE74C-37EE-46EA-BE2E-A6B28BB70D9A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RafaStore.Catalogo.API", "src\services\RafaStore.Catalogo.API\RafaStore.Catalogo.API.csproj", "{184DE74C-37EE-46EA-BE2E-A6B28BB70D9A}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{F2C3A6AD-8328-4CA9-A640-AF3269997C45}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RafaStore.WebAPI.Core", "RafaStore.WebAPI.Core\RafaStore.WebAPI.Core.csproj", "{A7003044-93A9-411A-B888-D3F6B7064118}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -49,6 +53,10 @@ Global
 		{184DE74C-37EE-46EA-BE2E-A6B28BB70D9A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{184DE74C-37EE-46EA-BE2E-A6B28BB70D9A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{184DE74C-37EE-46EA-BE2E-A6B28BB70D9A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A7003044-93A9-411A-B888-D3F6B7064118}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A7003044-93A9-411A-B888-D3F6B7064118}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A7003044-93A9-411A-B888-D3F6B7064118}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A7003044-93A9-411A-B888-D3F6B7064118}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -65,6 +73,8 @@ Global
 		{B3A8B714-DC52-496A-9BE1-06F8A7929B79} = {569F4418-F572-4D26-BFDC-99487C6B8D82}
 		{8C652B01-5A8B-4FB4-9C29-791A4ACC3105} = {F3D89C9F-73C9-4210-B499-4C76EED1A5A3}
 		{184DE74C-37EE-46EA-BE2E-A6B28BB70D9A} = {8C652B01-5A8B-4FB4-9C29-791A4ACC3105}
+		{F2C3A6AD-8328-4CA9-A640-AF3269997C45} = {A84A3880-7BE4-47E1-B142-32678B197C21}
+		{A7003044-93A9-411A-B888-D3F6B7064118} = {F2C3A6AD-8328-4CA9-A640-AF3269997C45}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8462BED5-3FC1-4AE9-BFF1-0F63839B3F3A}
