+++ src/services/RafaStore.Catalogo.API/RafaStore.Catalogo.API.csproj
@@ -19,6 +19,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<ProjectReference Include="..\..\..\RafaStore.WebAPI.Core\RafaStore.WebAPI.Core.csproj" />
 		<ProjectReference Include="..\..\building blocks\RafaStore.Core\RafaStore.Core.csproj" />
 	</ItemGroup>
