+++ src/services/RafaStore.Catalogo.API/appsettings.Development.json
@@ -7,5 +7,11 @@
   },
   "ConnectionStrings": {
     "DefaultConnection": "Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=C:\\Users\\rafa1\\OneDrive\\Documentos\\RafaStore.mdf;Integrated Security=True;Connect Timeout=30"
+  },
+  "AppSettings": {
+    "Secret": "RAFASTOREMEUSEGREDOSUPERSECRETODEMAIS",
+    "ExpiracaoHoras": 2,
+    "Emissor": "MeuSistema",
+    "ValidoEm": "https://localhost"
   }
 }
