+++ src/services/RafaStore.Identidade.API/Configuration/IdentityConfig.cs
@@ -1,16 +1,14 @@
-﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
 using RafaStore.Identidade.API.Data;
 using RafaStore.Identidade.API.Extensions;
-using System.Text;
+using RafaStore.WebAPI.Core.Identidade;
 
 namespace RafaStore.Identidade.API.Configuration;
 
 public static class IdentityConfig
 {
-    public static IServiceCollection AddIdentityConfiguration(this IServiceCollection services,
+    public static void AddIdentityConfiguration(this IServiceCollection services,
                                                               IConfiguration configuration)
     {
         services.AddDbContext<ApplicationDbContext>(options =>
@@ -22,38 +20,6 @@ public static class IdentityConfig
             .AddEntityFrameworkStores<ApplicationDbContext>()
             .AddDefaultTokenProviders();
 
-        var appSettingsSection = configuration.GetSection("AppSettings");
-        services.Configure<AppSettings>(appSettingsSection);
-
-        var appSettings = appSettingsSection.Get<AppSettings>();
-        var key = Encoding.ASCII.GetBytes(appSettings.Secret);
-
-        services.AddAuthentication(options =>
-        {
-            options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-            options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-        }).AddJwtBearer(options =>
-        {
-            options.RequireHttpsMetadata = true;
-            options.SaveToken = true;
-            options.TokenValidationParameters = new TokenValidationParameters
-            {
-                ValidateIssuerSigningKey = true,
-                IssuerSigningKey = new SymmetricSecurityKey(key),
-                ValidateIssuer = true,
-                ValidateAudience = true,
-                ValidAudience = appSettings.ValidoEm,
-                ValidIssuer = appSettings.Emissor
-            };
-        });
-
-        return services;
-    }
-
-    public static IApplicationBuilder UseIdentityConfiguration(this IApplicationBuilder app)
-    {
-        app.UseAuthorization();
-
-        return app;
+        services.AddJwtConfiguration(configuration);
     }
 }
