+++ src/services/RafaStore.Identidade.API/RafaStore.Identidade.API.csproj
@@ -25,6 +25,10 @@
 	</ItemGroup>
 
 	<ItemGroup>
+	  <ProjectReference Include="..\..\..\RafaStore.WebAPI.Core\RafaStore.WebAPI.Core.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
 	  <Compile Update="ErrosResources.Designer.cs">
 	    <DesignTime>True</DesignTime>
 	    <AutoGen>True</AutoGen>
